feat: Implement post search functionality and enhance post management

+++ backend/src/main/java/com/amalitech/communityboard/dto/PostResponse.java
@@ -14,5 +14,5 @@ public class PostResponse {
     private String authorEmail;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private int commentCount;
+    private Long commentCount;
 }
